# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cake Shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /promotions:
    get:
      operationId: listPromotions
      tags: [promotions]
      summary: List promotions
      responses:
        "200":
          description: List of promotions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Promotion"
    post:
      operationId: createPromotion
      tags: [promotions]
      summary: Create a promotion (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
  /promotions/{id}:
    patch:
      operationId: updatePromotion
      tags: [promotions]
      summary: Update a promotion (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
    delete:
      operationId: deletePromotion
      tags: [promotions]
      summary: Delete a promotion (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Promotion:
      type: object
      required: [id, title, active, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        code:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        bannerUrl:
          type: string
          nullable: true
        discountPct:
          type: number
          nullable: true
        discountAmount:
          type: number
          nullable: true
        active:
          type: boolean
        startsAt:
          type: string
          nullable: true
        endsAt:
          type: string
          nullable: true
        minimumOrderAmount:
          type: number
          nullable: true
        applicableCakeSlugs:
          type: array
          items:
            type: string
          nullable: true
        showInStrip:
          type: boolean
        createdAt:
          type: string
    PromotionInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        code:
          type: string
          nullable: true
        description:
          type: string
        bannerUrl:
          type: string
          nullable: true
        discountPct:
          type: number
          nullable: true
        discountAmount:
          type: number
          nullable: true
        active:
          type: boolean
        startsAt:
          type: string
          nullable: true
        endsAt:
          type: string
          nullable: true
        minimumOrderAmount:
          type: number
          nullable: true
        applicableCakeSlugs:
          type: array
          items:
            type: string
          nullable: true
        showInStrip:
          type: boolean
    PromotionUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        code:
          type: string
          nullable: true
        description:
          type: string
        bannerUrl:
          type: string
          nullable: true
        discountPct:
          type: number
          nullable: true
        discountAmount:
          type: number
          nullable: true
        active:
          type: boolean
        startsAt:
          type: string
          nullable: true
        endsAt:
          type: string
          nullable: true
        minimumOrderAmount:
          type: number
          nullable: true
        applicableCakeSlugs:
          type: array
          items:
            type: string
          nullable: true
        showInStrip:
          type: boolean